refactor: audit suggestions

+++ contracts/crowdfund/PartyBuy.sol
@@ -29,7 +29,7 @@ contract PartyBuy is PartyBuyBase {
         uint40 duration;
         // Maximum amount this crowdfund will pay for the NFT.
         // If zero, no maximum.
-        uint128 maximumPrice;
+        uint96 maximumPrice;
         // An address that receieves a portion of the final voting power
         // when the party transitions into governance.
         address payable splitRecipient;
@@ -85,7 +85,7 @@ contract PartyBuy is PartyBuyBase {
     ///         if it successfully buys the NFT.
     function buy(
         address payable callTarget,
-        uint128 callValue,
+        uint96 callValue,
         bytes calldata callData,
         FixedGovernanceOpts memory governanceOpts
     )
