+++ sol-tests/crowdfund/PartyBid.t.sol
@@ -49,7 +49,7 @@ contract PartyBidTest is Test, TestUtils {
     string defaultName = 'PartyBid';
     string defaultSymbol = 'PBID';
     uint40 defaultDuration = 60 * 60;
-    uint128 defaultMaxBid = 10e18;
+    uint96 defaultMaxBid = 10e18;
     address payable defaultSplitRecipient = payable(0);
     uint16 defaultSplitBps = 0.1e4;
     address defaultInitialDelegate;
@@ -74,7 +74,7 @@ contract PartyBidTest is Test, TestUtils {
     function _createCrowdfund(
         uint256 auctionId,
         uint256 tokenId,
-        uint128 initialContribution
+        uint96 initialContribution
     )
         private
         returns (PartyBid pb)
